+++ packages/terra-worklist-data-grid/src/subcomponents/Cell.jsx
@@ -142,7 +142,7 @@ function Cell(props) {
   const onMouseDown = ((event) => {
     if (!isFocusTrapEnabled) {
       onCellSelect({
-        rowId, columnId, rowIndex, columnIndex, isShiftKeyPressed: event.shiftKey, isCellSelectable: (!isMasked && isSelectable),
+        rowId, columnId, rowIndex, columnIndex, isShiftPressed: event.shiftKey, isCellSelectable: (!isMasked && isSelectable),
       });
     }
   });
@@ -176,7 +176,7 @@ function Cell(props) {
         case KeyCode.KEY_SPACE:
           if (onCellSelect) {
             onCellSelect({
-              rowId, columnId, rowIndex, columnIndex, isShiftKeyPressed: event.shiftKey, isCellSelectable: (!isMasked && isSelectable),
+              rowId, columnId, rowIndex, columnIndex, isShiftPressed: event.shiftKey, isCellSelectable: (!isMasked && isSelectable),
             });
           }
           event.preventDefault(); // prevent the default scrolling