+++ packages/terra-worklist-data-grid/src/subcomponents/Row.jsx
@@ -98,7 +98,7 @@ function Row(props) {
   const columnIndexOffSet = hasRowSelection ? 1 : 0;
 
   const handleCellSelect = useCallback((selectionDetails) => {
-    if (hasRowSelection || selectionDetails.isShiftKeyPressed) {
+    if (hasRowSelection || selectionDetails.isShiftPressed) {
       if (onRowSelect) {
         onRowSelect(selectionDetails);
       }