+++ packages/terra-worklist-data-grid/tests/jest/WorklistDataGrid.test.jsx
@@ -163,7 +163,7 @@ describe('basic grid', () => {
     maskedCell.at(0).simulate('keydown', { keyCode: 32 });
 
     // Validate mock function was called from simulated click event
-    expect(mockRowSelect).toHaveBeenCalledWith(['1'], []);
+    expect(mockRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }]);
 
     expect(wrapper).toMatchSnapshot();
   });
@@ -214,7 +214,7 @@ describe('basic grid', () => {
     nonSelectableCell.at(0).simulate('keydown', { keyCode: 32 });
 
     // Validate mock function was called from simulated click event
-    expect(mockRowSelect).toHaveBeenCalledWith(['1'], []);
+    expect(mockRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }]);
 
     expect(wrapper).toMatchSnapshot();
   });
@@ -339,11 +339,17 @@ describe('Error handling - prop types', () => {
 });
 
 describe('Row selection', () => {
-  it('verifies only onRowSelect is called when space is used to select a row.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
+  let mockOnCellSelect = null;
+  let mockOnRowSelect = null;
+  let mockOnEnableRowSelection = null;
 
+  beforeEach(() => {
+    mockOnCellSelect = jest.fn();
+    mockOnRowSelect = jest.fn();
+    mockOnEnableRowSelection = jest.fn();
+  });
+
+  it('verifies only onRowSelect is called when space is used to select a row.', () => {
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -362,7 +368,7 @@ describe('Row selection', () => {
     selectableCell.at(0).simulate('keydown', { keyCode: 32 });
 
     // Validate that only the onRowSelect callback is called.
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '3', selected: true }]);
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
 
@@ -370,10 +376,6 @@ describe('Row selection', () => {
   });
 
   it('verifies only onRowSelect is called when mouse is used to select a row.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -392,7 +394,7 @@ describe('Row selection', () => {
     selectableCell.at(0).simulate('mouseDown');
 
     // Validate that only the onRowSelect callback is called.
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '3', selected: true }]);
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
 
@@ -400,10 +402,6 @@ describe('Row selection', () => {
   });
 
   it('verifies only onRowSelect is called when space is used to unselect a row.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -422,7 +420,7 @@ describe('Row selection', () => {
     selectableCell.at(0).simulate('keydown', { keyCode: 32 });
 
     // Validate that only the onRowSelect callback is called.
-    expect(mockOnRowSelect).toHaveBeenCalledWith([], ['4']);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '4', selected: false }]);
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
 
@@ -430,10 +428,6 @@ describe('Row selection', () => {
   });
 
   it('verifies only onRowSelect is called when mouse is used to unselect a row.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -452,7 +446,7 @@ describe('Row selection', () => {
     selectableCell.at(0).simulate('mouseDown');
 
     // Validate that only the onRowSelect callback is called.
-    expect(mockOnRowSelect).toHaveBeenCalledWith([], ['4']);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '4', selected: false }]);
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
 
@@ -460,10 +454,6 @@ describe('Row selection', () => {
   });
 
   it('verifies callbacks when Shift+Down is used and row selection mode is not enabled.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -481,17 +471,13 @@ describe('Row selection', () => {
     expect(mockOnCellSelect).toHaveBeenCalledWith('1', 'Column-1'); // The first click to select the cell from which shift+Down will occur.
 
     selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }]);
     expect(mockOnEnableRowSelection).toHaveBeenCalled();
 
     expect(wrapper).toMatchSnapshot();
   });
 
   it('verifies callbacks when Shift+Down is used and row selection mode is enabled.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -508,7 +494,7 @@ describe('Row selection', () => {
     const selectableCell = wrapper.find(Row).at(0).find('td.selectable');
     selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
 
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }]);
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
 
@@ -516,10 +502,6 @@ describe('Row selection', () => {
   });
 
   it('verifies callbacks when Shift+Down more than one.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -537,8 +519,8 @@ describe('Row selection', () => {
     selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
     selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
 
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []);
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }]);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }, { id: '3', selected: true }]);
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
 
@@ -546,10 +528,6 @@ describe('Row selection', () => {
   });
 
   it('verifies multiple independent ranges can be created with Shift+Down.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -566,7 +544,7 @@ describe('Row selection', () => {
     // Find a cell on row 1 and use Shift+Down to create a selected range of rows.
     let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
     selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }]);
 
     // Release the Shift key and Arrow Down once to get to row 3.
     selectableCell.at(0).simulate('keyup', { keyCode: 16 });
@@ -575,7 +553,7 @@ describe('Row selection', () => {
     // Find a cell on Row 3 and do a Shift+Down
     selectableCell = wrapper.find(Row).at(2).find('td.selectable');
     selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['3', '4'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '3', selected: true }, { id: '4', selected: true }]);
 
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
@@ -584,10 +562,6 @@ describe('Row selection', () => {
   });
 
   it('verifies callbacks when Shift+Down contracts the range of selected rows.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -606,9 +580,9 @@ describe('Row selection', () => {
     selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 }); // Shift+Down extends the range
     selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 38 }); // Shift+Up contracts the range
 
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []); // range is established with 2 rows selected.
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []); // range is expanded with an additional row selected.
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], ['3']); // range is contracted with with one row unselected.
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }]); // range is established with 2 rows selected.
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }, { id: '3', selected: true }]); // range is expanded with an additional row selected.
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }, { id: '3', selected: false }]); // range is contracted with with one row unselected.
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
 
@@ -616,10 +590,6 @@ describe('Row selection', () => {
   });
 
   it('verifies Shift+Click honors the anchor established by row selection using space bar.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -636,12 +606,12 @@ describe('Row selection', () => {
     let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
     // Select Row 1 using space.
     selectableCell.at(0).simulate('keydown', { keyCode: 32 });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }]);
 
     // Shift+Click on row 3
     selectableCell = wrapper.find(Row).at(2).find('td.selectable');
     selectableCell.at(0).simulate('mouseDown', { shiftKey: true });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }, { id: '3', selected: true }]);
 
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
@@ -650,10 +620,6 @@ describe('Row selection', () => {
   });
 
   it('verifies Shift+Space honors the anchor established by row selection using Mouse.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -670,12 +636,12 @@ describe('Row selection', () => {
     let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
     // Select Row 1 using Mouse.
     selectableCell.at(0).simulate('mouseDown');
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }]);
 
     // Shift+Space on row 3
     selectableCell = wrapper.find(Row).at(2).find('td.selectable');
     selectableCell.at(0).simulate('keyDown', { shiftKey: true, keyCode: 32 });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }, { id: '3', selected: true }]);
 
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
@@ -684,10 +650,6 @@ describe('Row selection', () => {
   });
 
   it('verifies Shift+Click enables row selection when row selection is turned off.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -703,7 +665,7 @@ describe('Row selection', () => {
     const selectableCell = wrapper.find(Row).at(2).find('td.selectable');
     // Select Row 3 using Shift+Click when row selection is off.
     selectableCell.at(0).simulate('mouseDown', { shiftKey: true });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '3', selected: true }]);
     expect(mockOnEnableRowSelection).toHaveBeenCalled();
 
     expect(mockOnCellSelect).not.toHaveBeenCalled();
@@ -712,10 +674,6 @@ describe('Row selection', () => {
   });
 
   it('verifies Shift+Space enables row selection when row selection is turned off.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -731,7 +689,7 @@ describe('Row selection', () => {
     const selectableCell = wrapper.find(Row).at(2).find('td.selectable');
     // Select Row 3 using Shift+Space when row selection is off.
     selectableCell.at(0).simulate('keyDown', { shiftKey: true, keyCode: 32 });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '3', selected: true }]);
     expect(mockOnEnableRowSelection).toHaveBeenCalled();
 
     expect(mockOnCellSelect).not.toHaveBeenCalled();
@@ -740,10 +698,6 @@ describe('Row selection', () => {
   });
 
   it('verifies Shift+Click establishes anchor and selects row when row selection is already enabled.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -760,12 +714,12 @@ describe('Row selection', () => {
     let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
     // Select Row 1 using Shift+Click. Row is selected and anchor created.
     selectableCell.at(0).simulate('mouseDown', { shiftKey: true });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }]);
 
     // Shift+Click on row 3
     selectableCell = wrapper.find(Row).at(2).find('td.selectable');
     selectableCell.at(0).simulate('mouseDown', { shiftKey: true });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }, { id: '3', selected: true }]);
 
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();
@@ -774,10 +728,6 @@ describe('Row selection', () => {
   });
 
   it('verifies Shift+Space establishes anchor and selects row if row selection is already enabled.', () => {
-    const mockOnCellSelect = jest.fn();
-    const mockOnRowSelect = jest.fn();
-    const mockOnEnableRowSelection = jest.fn();
-
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
@@ -794,12 +744,12 @@ describe('Row selection', () => {
     let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
     // Select Row 1 using Shift+Space. Row is selected and anchor created.
     selectableCell.at(0).simulate('keyDown', { shiftKey: true, keyCode: 32 });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }]);
 
     // Shift+Space on row 3
     selectableCell = wrapper.find(Row).at(2).find('td.selectable');
     selectableCell.at(0).simulate('keyDown', { shiftKey: true, keyCode: 32 });
-    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith([{ id: '1', selected: true }, { id: '2', selected: true }, { id: '3', selected: true }]);
 
     expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
     expect(mockOnCellSelect).not.toHaveBeenCalled();